Revert "Tenant provisioning in the dataplane (#2590)"
This reverts commit 87eb4add5d6c555d3a19fc808980a82a3fb022ab.
